* Inline Function again * Move the job of declaring variables on the scope to the individual   Identifiers * Extract Variable types and refactor * In order to find out if a variable has an effect when mutated, ask it! * Now var statements should be correctly associated (test pending)
