+++ src/ast/nodes/ArrowFunctionExpression.js
@@ -1,7 +1,18 @@
-import Function from './shared/Function.js';
+import Node from '../Node';
 import Scope from '../scopes/Scope.js';
 
-export default class ArrowFunctionExpression extends Function {
+export default class ArrowFunctionExpression extends Node {
+	hasEffects () {
+		return this.included;
+	}
+
+	initialiseChildren () {
+		this.params.forEach( param => param.initialiseAndDeclare( this.scope, 'parameter' ) );
+		this.body.initialiseAndReplaceScope ?
+			this.body.initialiseAndReplaceScope( this.scope ) :
+			this.body.initialise( this.scope );
+	}
+
 	initialiseScope ( parentScope ) {
 		this.scope = new Scope( {
 			parent: parentScope,
