+++ src/ast/nodes/ClassExpression.js
@@ -2,7 +2,7 @@ import Class from './shared/Class.js';
 
 export default class ClassExpression extends Class {
 	initialiseChildren ( parentScope ) {
-		this.id && this.id.initialiseAndDeclare( this.scope, 'let', this );
+		this.id && this.id.initialiseAndDeclare( this.scope, 'class', this );
 		super.initialiseChildren( parentScope );
 	}
 }
