+++ src/ast/nodes/FunctionExpression.js
@@ -1,8 +1,24 @@
-import Function from './shared/Function.js';
+import Node from '../Node.js';
+import Scope from '../scopes/Scope.js';
 
-export default class FunctionExpression extends Function {
-	initialiseChildren ( parentScope ) {
-		this.id && this.id.initialiseAndDeclare( this.scope, 'var', this );
-		super.initialiseChildren( parentScope );
+export default class FunctionExpression extends Node {
+	hasEffects () {
+		return this.included || (this.id && this.id.hasEffects());
+	}
+
+	initialiseChildren () {
+		this.id && this.id.initialiseAndDeclare( this.scope, 'function', this );
+		this.params.forEach( param => param.initialiseAndDeclare( this.scope, 'parameter' ) );
+		this.body.initialiseAndReplaceScope ?
+			this.body.initialiseAndReplaceScope( this.scope ) :
+			this.body.initialise( this.scope );
+	}
+
+	initialiseScope ( parentScope ) {
+		this.scope = new Scope( {
+			parent: parentScope,
+			isBlockScope: false,
+			isLexicalBoundary: true
+		} );
 	}
 }
