+++ src/ast/nodes/Property.js
@@ -10,9 +10,9 @@ export default class Property extends Node {
 		return this.value.hasEffectsWhenAssigned( options );
 	}
 
-	initialiseAndDeclare ( parentScope, kind ) {
+	initialiseAndDeclare ( parentScope, kind, init ) {
 		this.initialiseScope( parentScope );
-		this.value.initialiseAndDeclare( parentScope, kind, UNKNOWN_ASSIGNMENT );
+		this.value.initialiseAndDeclare( parentScope, kind, init && UNKNOWN_ASSIGNMENT );
 	}
 
 	render ( code, es ) {
