+++ src/ast/variables/NamespaceVariable.js
@@ -1,0 +1,54 @@
+import Variable from './Variable';
+import { blank, forOwn, keys } from '../../utils/object';
+import { reservedWords } from '../../utils/identifierHelpers.js';
+import { UNKNOWN_ASSIGNMENT } from '../values';
+
+export default class NamespaceVariable extends Variable {
+	constructor ( module ) {
+		super( module.basename() );
+		this.isNamespace = true;
+		this.module = module;
+		this.needsNamespaceBlock = false;
+
+		this.originals = blank();
+		module.getExports().concat( module.getReexports() ).forEach( name => {
+			this.originals[ name ] = module.traceExport( name );
+		} );
+	}
+
+	addReference ( node ) {
+		this.name = node.name;
+	}
+
+	assignExpression () {}
+
+	gatherPossibleValues ( values ) {
+		values.add( UNKNOWN_ASSIGNMENT );
+	}
+
+	includeDeclaration () {
+		if ( this.included ) {
+			return false;
+		}
+		this.included = true;
+		this.needsNamespaceBlock = true;
+		forOwn( this.originals, original => original.includeDeclaration() );
+		return true;
+	}
+
+	renderBlock ( es, legacy, indentString ) {
+		const members = keys( this.originals ).map( name => {
+			const original = this.originals[ name ];
+
+			if ( original.isReassigned && !legacy ) {
+				return `${indentString}get ${name} () { return ${original.getName( es )}; }`;
+			}
+
+			if ( legacy && ~reservedWords.indexOf( name ) ) name = `'${name}'`;
+			return `${indentString}${name}: ${original.getName( es )}`;
+		} );
+
+		const callee = legacy ? `(Object.freeze || Object)` : `Object.freeze`;
+		return `${this.module.bundle.varOrConst} ${this.getName( es )} = ${callee}({\n${members.join( ',\n' )}\n});\n\n`;
+	}
+}
